% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ak91.R
\docType{data}
\name{ak91}
\alias{ak91}
\title{Quarter of Birth and Return to Schooling Data}
\format{A data frame.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/02/ak91.dta}
}
\usage{
ak91
}
\description{
Data from Angrist and Krueger (1991), which estimate returns to schooling
using quarters of birth as an instrument for years of schooling.
These data are used in Tables 6.4. and 6.5 and Figures 6.1 and 6.2 of \emph{Mastering 'Metrics}.
}
\references{
\itemize{
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/02/ReadMe_QOB.txt}
\item \url{http://masteringmetrics.com/resources/}
\item Angrist, J. D., and Krueger, A. B. (1991) \dQuote{Does Compulsory School Attendance Affect Schooling and Earnings}, \emph{Quarterly Journal of Economics} \url{}
}
}
\keyword{datasets}
