% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks.R
\docType{data}
\name{banks}
\alias{banks}
\title{Number of Mississippi Banks, 1929--1933}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1878 rows and 5 columns.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/02/banks.csv}
}
\usage{
banks
}
\description{
The number of banks in Mississippi in the 6th and 9th
Federal Districts from 1929--1933. This is from
Richardson and Troost (2009) and used in Ch. 5 of
\emph{Mastering `Metrics}.
}
\references{
\itemize{
\item Angrist, J., Pischke, J.-S. (2014) \dQuote{Mastering 'Metrics: The Path from Cause to Effect}.
\item Richardson, G., Troost, W. (2009) \dQuote{Monetary Intervention Mitigated Banking Panics during the Great Depression: Quasi-Experimental Evidence from a Federal Reserve District Border, 1929-1933}, \emph{Journal of Political Economy} \href{https://doi.org/10.1086/649603}{DOI:10.1086/649603}
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/01/ReadMe_MLDA_DD.txt}
}
}
\keyword{datasets}
