% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_labor.R
\docType{data}
\name{child_labor}
\alias{child_labor}
\title{Child Labor Law Data}
\format{A data frame.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/02/AA_small.dta_.zip}
}
\usage{
child_labor
}
\description{
Data from Acemoglu and Angrist (2000) which used 2SLS to
estimate returns to schooling using child labor laws as instruments for
years of schooling. These data is used in Table 6.3 of
\emph{Mastering 'Metrics}.
}
\references{
\itemize{
\item Acemoglu, D., Angrist, J. (2000) \dQuote{How Large Are Human-Capital Externalities? Evidence From Compulsory Schooling Laws}, \emph{NBER Macroeconomics Annual} \href{https://doi.org/10.1086/654403}{DOI:10.1086/654403}
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/02/ReadMe_ChildLaborLaws.txt}
}
}
\keyword{datasets}
