% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\docType{data}
\name{deaths}
\alias{deaths}
\title{State}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24786 rows and 15 columns.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/01/deaths.dta}
}
\usage{
deaths
}
\description{
State
}
\references{
\itemize{
\item Angrist, J., Pischke, J.-S. (2014) \dQuote{Mastering 'Metrics: The Path from Cause to Effect}.
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/01/ReadMe_MLDA_DD.txt}
\item Du Mouchel, W., Williams, A. F.,  Zador, P. (1987) \dQuote{Raising the Alcohol Purchase Age: Its Effects on Fatal Motor Vehicle Crashes in Twenty-Six States}, \emph{The Journal of Legal Studies}, \url{http://www.jstor.org/stable/724480?origin=JSTOR-pdf}.
\item Norberg, K. E., Bierut, L. J., Grucza, R. A. (2009) \dQuote{Long-Term Effects of Minimum Drinking Age Laws on Past-Year Alcohol and Drug Use Disorders}, \emph{Alcoholism: Clinical and Experimental Research} \href{https://dx.doi.org/10.1111/j.1530-0277.2009.01056.x}{DOI:10.1111/j.1530-0277.2009.01056.x}
}
}
\keyword{datasets}
