% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masteringmetrics.R
\docType{package}
\name{masteringmetrics}
\alias{masteringmetrics}
\alias{masteringmetrics-package}
\title{masteringmetrics: A package with data from Mastering 'Metrics book}
\description{
The \pkg{masteringmetrics} package contains data used in the empirical examples of Angrist and Pischke \emph{Mastering 'Metrics}.
These datasets and replication code are available from \url{http://masteringmetrics.com/resources/}.
}
\details{
Installing this package also installs packages that the R code replicating \emph{Mastering 'Metrics} that is available from \url{https://jrnold.github.io/masteringmetrics}.
}
