% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdve.R
\docType{data}
\name{mdve}
\alias{mdve}
\title{Minneapolis Domestic Violence Experiment (MDVE)}
\format{A data frame.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/02/mdve.dta}
}
\usage{
mdve
}
\description{
Data from the Minneapolis Domestic Violence Experiment (MDVE)
used in Table 3.3 of \emph{Mastering 'Metrics}.
}
\references{
\itemize{
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/02/ReadMe_MDVE.txt}
\item Angrist, J. (2006) \dQuote{Instrumental variables methods in experimental criminological research: what, why and how}.
\emph{Journal of Experimental Criminology} \href{https://dx.doi.org/10.1007/s11292-005-5126-x}{DOI:10.1007/s11292-005-5126-x}
\item Sherman, L.A., Berk, R.A. (1984)
\dQuote{The Specific Deterrent Effects of Arrest for Domestic Assault}
\emph{American Sociological Review}
}
}
\keyword{datasets}
