% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlda.R
\docType{data}
\name{mlda}
\alias{mlda}
\title{Minimum Legal Drinking Age Age-Fatalities Data}
\format{A data frame.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/01/AEJfigs.dta}
}
\usage{
mlda
}
\description{
Data on mortality rates by age-groups for minimum legal drinking ages
the regression discontinuity analysis in
Table 4.1 and Figures 4.2, 4.4, and 4.5 in \emph{Mastering 'Metrics}.
Table 4.1 presents sharp RD estimates of the effect of the minimum legal drinking age (MLDA) on mortality.
}
\references{
\itemize{
\item Carpenter, C., Dobkin, C. (2011) \dQuote{The Minimum Legal Drinking Age and Public Health}
\emph{Journal of Economic Perspectives}.
\item Carpenter, C., Dobkin, C. \dQuote{The Effect of Alcohol Consumption on Mortality: Regression Discontinuity Evidence from the Minimum Drinking Age}
\emph{American Economic Review}.
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/01/ReadMe_MLDA.txt}
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/01/master_cd_rd.do}
}
}
\keyword{datasets}
