% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHIS2009.R
\docType{data}
\name{NHIS2009}
\alias{NHIS2009}
\title{2009 National Health Interview Survey (NHIS)}
\format{A data frame.}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/01/Data.zip}
}
\usage{
NHIS2009
}
\description{
Data from the 2009 National Health Interview Survey (NHIS).

Angrist, J. D., Pischke, J.-S. (2014) *Mastering 'Metrics: The Path from Cause to Effect*.
}
\details{
This is used in Table 1.1 of \emph{Mastering 'Metrics}.
}
\references{
\itemize{
\item \url{https://nhis.ipums.org/nhis/}
}
}
\keyword{datasets}
