% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubtwins.R
\docType{data}
\name{pubtwins}
\alias{pubtwins}
\title{Twinsburg Twins Returns to Schooling Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 680 rows and 39 columns.}
\source{
\url{http://arks.princeton.edu/ark:/88435/dsp01rv042t084}
}
\usage{
pubtwins
}
\description{
Returns to schooling for the Twinsburg Twisn used in
Ashenfelter and Krueger (1994) and Ashenfelter and Rouse (1998).
}
\references{
\itemize{
\item Ashenfelter, O., Krueger, A. (1994) "Estimates of the Economic Return to Schooling from a New Sample of Twins", \emph{American Economic Review} \url{http://www.jstor.org/stable/2117766}
\item Ashenfelter, O., Krueger, A. (1998) "Income, Schooling, and Ability: Evidence from a New Sample of Identical Twins", \emph{Quarterly Journal of Economics} \href{https://doi.org/10.1162/003355398555577}{DOI:10.1162/003355398555577}
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/02/banks.csv}
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/02/ReadMe_Twinsburg.txt}
}
}
\keyword{datasets}
