% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_person_spend.R
\docType{data}
\name{rand_person_spend}
\alias{rand_person_spend}
\alias{rand_sample}
\title{RAND Health Insurance Experiment (HIE) Data}
\format{\itemize{
\item \code{rand_person_spend}: A data frame.
\item \code{rand_sample}: A data frame.
}}
\source{
\url{http://masteringmetrics.com/wp-content/uploads/2015/01/Data1.zip}
}
\usage{
rand_person_spend

rand_sample
}
\description{
The data frame \code{rand_sample} presents demographic and
baseline health characteristics for subjects of the
RAND Health Insurance Experiment (HIE).
This was generated from \code{rand_initial_sample_2.dta}.
}
\details{
The data frame \code{rand_person_spend} reports health outcome and
health expenditure data from the RAND Health Insurance Experiment (HIE).
This was generated from \code{person_years.dta} and \code{annual_spend.dta}.
}
\references{
\itemize{
\item \url{http://masteringmetrics.com/wp-content/uploads/2015/01/ReadMe_RAND.txt}
\item Aron-Dine, A., Einav, L., Finkelstein, A. (2013) \dQuote{The RAND Health Insurance Experiment, Three Decades Later}
\emph{Journal of Economic Perspectives}
}
}
\keyword{datasets}
